resourses/js配下のjsファイル整理。location.pathnameの値でどのモジュールを使用するか振り分ける。
