+++ src/public/js/app.js
@@ -4965,18 +4965,26 @@ module.exports = {
 "use strict";
 __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _home_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./home.js */ "./resources/js/home.js");
-/* harmony import */ var _home_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_home_js__WEBPACK_IMPORTED_MODULE_0__);
 /* harmony import */ var _tasks_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./tasks.js */ "./resources/js/tasks.js");
-/* harmony import */ var _tasks_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_tasks_js__WEBPACK_IMPORTED_MODULE_1__);
 /* harmony import */ var alpinejs__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! alpinejs */ "./node_modules/alpinejs/dist/module.esm.js");
+__webpack_require__(/*! jquery-ui/ui/widgets/draggable */ "./node_modules/jquery-ui/ui/widgets/draggable.js");
+
+__webpack_require__(/*! jquery-ui/ui/widgets/droppable */ "./node_modules/jquery-ui/ui/widgets/droppable.js");
+
 __webpack_require__(/*! ./bootstrap */ "./resources/js/bootstrap.js");
 
 
 
 
-__webpack_require__(/*! jquery-ui/ui/widgets/draggable */ "./node_modules/jquery-ui/ui/widgets/draggable.js");
+window.onload = function () {
+  var path = location.pathname;
 
-__webpack_require__(/*! jquery-ui/ui/widgets/droppable */ "./node_modules/jquery-ui/ui/widgets/droppable.js");
+  if (path === "/tasks") {
+    (0,_tasks_js__WEBPACK_IMPORTED_MODULE_1__.TaskEvent)();
+  } else if (path === "/") {
+    (0,_home_js__WEBPACK_IMPORTED_MODULE_0__.HomeEvent)();
+  }
+};
 
 
 window.Alpine = alpinejs__WEBPACK_IMPORTED_MODULE_2__["default"];
@@ -5019,35 +5027,110 @@ window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 /*!******************************!*\
   !*** ./resources/js/home.js ***!
   \******************************/
-/***/ (() => {
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-window.onload = onClick;
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "HomeEvent": () => (/* binding */ HomeEvent)
+/* harmony export */ });
+/* provided dependency */ var $ = __webpack_require__(/*! jquery */ "./node_modules/jquery/dist/jquery.js");
+function _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== "undefined" && o[Symbol.iterator] || o["@@iterator"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === "number") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it["return"] != null) it["return"](); } finally { if (didErr) throw err; } } }; }
 
-function onClick() {
+function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }
+
+function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }
+
+function HomeEvent() {
   var button = document.querySelector('#time-button');
-  var timer = document.querySelector('#setTimer');
-  console.log('gda');
-
-  if (button === null) {
-    return false;
-  }
-
+  var timer = document.querySelectorAll('#setTimer');
+  var timerDisplay = document.querySelector('#timer-display');
+  var clickFlag = true;
   button.addEventListener('click', function () {
-    var spin = document.querySelector('#spin');
-    spin.classList.toggle('animate-spin');
+    if (clickFlag) {
+      var amount = timerDisplay.getAttribute('data-timer-amount');
+      var now = new Date();
+      var end = new Date(now.getTime() + amount * 1000);
+      var time = end.getTime(); // 残りtimestamp
+
+      var intervalId = setInterval(function () {
+        time -= 1000;
+        var remain = time - now.getTime();
+        timerDisplay.setAttribute('data-timer-amount', remain);
+        var hour = Math.floor(remain / (60 * 1000));
+        var min = Math.floor(remain % (60 * 1000) / 1000);
+        timerDisplay.textContent = ('0' + hour).slice(-2) + ':' + ('0' + min).slice(-2);
+
+        if (remain < 1) {
+          clearInterval(intervalId);
+          document.querySelector('#finish-icon').classList.remove('hidden');
+
+          var _finishIntervalId = setTimeout(function () {
+            document.querySelector('#finish-icon').classList.add('hidden');
+          }, 5000); // のちにここに何回達成したかを更新する同期処理を追加
+
+
+          location.reload();
+        }
+
+        var _iterator = _createForOfIteratorHelper(timer),
+            _step;
+
+        try {
+          for (_iterator.s(); !(_step = _iterator.n()).done;) {
+            var t = _step.value;
+            t.addEventListener('click', function () {
+              clearInterval(intervalId);
+              clearInterval(finishIntervalId);
+            });
+          }
+        } catch (err) {
+          _iterator.e(err);
+        } finally {
+          _iterator.f();
+        }
+
+        var deg = (1 - remain / amount / 1000) * 360;
+        console.log(deg);
+        $('#spin').css('transform', 'rotate(' + deg + 'deg)');
+      }, 1000);
+    }
+
+    clickFlag = false;
   }, false);
-  timer.addEventListener('click', function (event) {
-    var task_id = event.currentTarget.getAttribute('data-task-id');
-    var url = "/home/" + task_id;
-    var timer = document.querySelector('.timer');
-    fetch(url, {}).then(function (response) {
-      return response.json();
-    }).then(function (data) {
-      timer.textContent = data["timer"];
-    })["catch"](function (err) {
-      alert('通信に失敗しました。');
-    });
-  });
+
+  if (timer.length) {
+    var _iterator2 = _createForOfIteratorHelper(timer),
+        _step2;
+
+    try {
+      for (_iterator2.s(); !(_step2 = _iterator2.n()).done;) {
+        var t = _step2.value;
+        t.addEventListener('click', function (event) {
+          var task_id = event.currentTarget.getAttribute('data-tasks-id');
+          var url = "/home/" + task_id;
+          var timer = document.querySelector('.timer');
+          fetch(url, {}).then(function (response) {
+            return response.json();
+          }).then(function (data) {
+            if (data["timer"]) {
+              timerDisplay.textContent = ('0' + data["timer"]).slice(-2) + ':00';
+              timerDisplay.setAttribute('data-timer-amount', data["timer"] * 60);
+            } else {
+              timerDisplay.textContent = '60:00';
+              timerDisplay.setAttribute('data-timer-amount', '1800');
+            }
+          })["catch"](function (err) {
+            alert('通信に失敗しました。');
+          });
+        });
+      }
+    } catch (err) {
+      _iterator2.e(err);
+    } finally {
+      _iterator2.f();
+    }
+  }
 }
 
 /***/ }),
@@ -5056,8 +5139,13 @@ function onClick() {
 /*!*******************************!*\
   !*** ./resources/js/tasks.js ***!
   \*******************************/
-/***/ ((__unused_webpack_module, __unused_webpack_exports, __webpack_require__) => {
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "TaskEvent": () => (/* binding */ TaskEvent)
+/* harmony export */ });
 /* provided dependency */ var $ = __webpack_require__(/*! jquery */ "./node_modules/jquery/dist/jquery.js");
 function _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== "undefined" && o[Symbol.iterator] || o["@@iterator"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === "number") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it["return"] != null) it["return"](); } finally { if (didErr) throw err; } } }; }
 
@@ -5065,9 +5153,7 @@ function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o =
 
 function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }
 
-window.onload = Task;
-
-function Task() {
+function TaskEvent() {
   // Task
   var tasks = document.querySelectorAll('.task');
   var trashcan = document.querySelector('.fa-trash-can');
@@ -5202,101 +5288,6 @@ function Task() {
     exhaust.addEventListener('click', function () {
       location.href = "/tasks/forceDelete";
     });
-  } // 後で別ファイルで実行できるように修正する
-  // Top
-
-
-  var button = document.querySelector('#time-button');
-  var timer = document.querySelectorAll('#setTimer');
-  var timerDisplay = document.querySelector('#timer-display');
-  var clickFlag = true;
-
-  if (button !== null) {
-    button.addEventListener('click', function () {
-      if (clickFlag) {
-        var amount = timerDisplay.getAttribute('data-timer-amount');
-        var now = new Date();
-        var end = new Date(now.getTime() + amount * 1000);
-        var time = end.getTime(); // 残りtimestamp
-
-        var intervalId = setInterval(function () {
-          time -= 1000;
-          var remain = time - now.getTime();
-          timerDisplay.setAttribute('data-timer-amount', remain);
-          var hour = Math.floor(remain / (60 * 1000));
-          var min = Math.floor(remain % (60 * 1000) / 1000);
-          timerDisplay.textContent = ('0' + hour).slice(-2) + ':' + ('0' + min).slice(-2);
-
-          if (remain < 1) {
-            clearInterval(intervalId);
-            document.querySelector('#finish-icon').classList.remove('hidden');
-
-            var _finishIntervalId = setTimeout(function () {
-              document.querySelector('#finish-icon').classList.add('hidden');
-            }, 5000); // のちにここに何回達成したかを更新する同期処理を追加
-
-
-            location.reload();
-          }
-
-          var _iterator2 = _createForOfIteratorHelper(timer),
-              _step2;
-
-          try {
-            for (_iterator2.s(); !(_step2 = _iterator2.n()).done;) {
-              var t = _step2.value;
-              t.addEventListener('click', function () {
-                clearInterval(intervalId);
-                clearInterval(finishIntervalId);
-              });
-            }
-          } catch (err) {
-            _iterator2.e(err);
-          } finally {
-            _iterator2.f();
-          }
-
-          var deg = (1 - remain / amount / 1000) * 360;
-          console.log(deg);
-          $('#spin').css('transform', 'rotate(' + deg + 'deg)');
-        }, 1000);
-      }
-
-      clickFlag = false;
-    }, false);
-  }
-
-  if (timer.length) {
-    var _iterator3 = _createForOfIteratorHelper(timer),
-        _step3;
-
-    try {
-      for (_iterator3.s(); !(_step3 = _iterator3.n()).done;) {
-        var t = _step3.value;
-        t.addEventListener('click', function (event) {
-          var task_id = event.currentTarget.getAttribute('data-tasks-id');
-          var url = "/home/" + task_id;
-          var timer = document.querySelector('.timer');
-          fetch(url, {}).then(function (response) {
-            return response.json();
-          }).then(function (data) {
-            if (data["timer"]) {
-              timerDisplay.textContent = ('0' + data["timer"]).slice(-2) + ':00';
-              timerDisplay.setAttribute('data-timer-amount', data["timer"] * 60);
-            } else {
-              timerDisplay.textContent = '60:00';
-              timerDisplay.setAttribute('data-timer-amount', '1800');
-            }
-          })["catch"](function (err) {
-            alert('通信に失敗しました。');
-          });
-        });
-      }
-    } catch (err) {
-      _iterator3.e(err);
-    } finally {
-      _iterator3.f();
-    }
   }
 }
 
@@ -36785,18 +36776,6 @@ module.exports = JSON.parse('{"_from":"axios@^0.21","_id":"axios@0.21.4","_inBun
 /******/ 				}
 /******/ 			}
 /******/ 			return result;
-/******/ 		};
-/******/ 	})();
-/******/ 	
-/******/ 	/* webpack/runtime/compat get default export */
-/******/ 	(() => {
-/******/ 		// getDefaultExport function for compatibility with non-harmony modules
-/******/ 		__webpack_require__.n = (module) => {
-/******/ 			var getter = module && module.__esModule ?
-/******/ 				() => (module['default']) :
-/******/ 				() => (module);
-/******/ 			__webpack_require__.d(getter, { a: getter });
-/******/ 			return getter;
 /******/ 		};
 /******/ 	})();
 /******/
