+++ src/resources/js/app.js
@@ -1,9 +1,17 @@
-require('./bootstrap');
-import './home.js';
-import './tasks.js';
-
 require('jquery-ui/ui/widgets/draggable');
 require('jquery-ui/ui/widgets/droppable');
+require('./bootstrap');
+import { HomeEvent } from './home.js';
+import { TaskEvent } from './tasks.js';
+
+window.onload = () => {
+  let path = location.pathname;
+  if (path === "/tasks") {
+    TaskEvent();
+  } else if (path === "/") {
+    HomeEvent();
+  }
+};
 
 import Alpine from 'alpinejs';
